// frontend/app/src/app/models/pregled.ts
export class Pregled{
    lekar: string;
    naziv: string;
    trajanje: number[];
    pocetak: number[];
    cena: number;
    status: string;
    specijalizacija: string;
    datum: string;
    pacijent: string;
    ordinacija: string;
    komentar: string;
    izvestaj: boolean;
    obavesten: boolean;
}